feat(error): added new errors to controllers, services and middlewares

+++ src/services/posts/postDelete.service.ts
@@ -1,13 +1,18 @@
 import { AppDataSource } from "../../data-source";
 import { Project_Posts } from "../../entities/project_posts.entity";
+import { AppError } from "../../errors/appError";
 import { IPostId } from "../../interfaces/posts";
 
-const postDeleteService = async ({id}: IPostId) => {
+const postDeleteService = async ({ id }: IPostId) => {
   const postsRepository = AppDataSource.getRepository(Project_Posts);
 
   const posts = await postsRepository.find();
 
   const postDeleted = posts.find((posts) => posts.id === id);
+
+  if (!postDeleted) {
+    throw new AppError("Post not found", 404);
+  }
 
   await postsRepository.delete(postDeleted!.id);
