+++ src/services/projects/projectDelete.service.ts
@@ -1,5 +1,6 @@
 import { AppDataSource } from "../../data-source";
 import { Project } from "../../entities/project.entity";
+import { AppError } from "../../errors/appError";
 
 const projectDeleteSelfService = async (id: string) => {
   const projectRepository = AppDataSource.getRepository(Project);
@@ -7,6 +8,10 @@ const projectDeleteSelfService = async (id: string) => {
   const projects = await projectRepository.find();
 
   const projectDeleted = projects.find((project) => project.id === id);
+
+  if(!projectDeleted){
+    throw new AppError("Project not found", 404)
+  }
 
   await projectRepository.delete(projectDeleted!.id);
